# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ATLAS \u2014 Market Intelligence Engine",
    "description": "Jhaveri Intelligence Platform \u2014 Market \u2192 Sector \u2192 Stock \u2192 Decision",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/query": {
      "post": {
        "tags": [
          "query"
        ],
        "summary": "Execute a raw UQL query",
        "description": "Spec \u00a717.5\u201317.7. Accepts a full UQL envelope \u2014 snapshot, aggregation (`group_by` + `aggregations`), or `mode='timeseries'` with `time_range`/`granularity`. Every fixed endpoint under `/stocks` and `/mf` is a thin wrapper over this route, so any payload the engine accepts here is safe to ship.",
        "operationId": "execute_query_api_v1_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UQLRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Query executed \u2014 records + total + _meta",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UQLResponse"
                }
              }
            }
          },
          "400": {
            "description": "UQL validation rejection (safety, limits, shape)"
          },
          "503": {
            "description": "Entity partition missing (FR-019)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/query/template": {
      "post": {
        "tags": [
          "query"
        ],
        "summary": "Execute a named UQL template",
        "description": "Spec \u00a717.7. Dispatches a registered template (`backend/services/uql/templates.py::REGISTRY`) with optional params. Adding a template requires no route code.",
        "operationId": "execute_query_template_api_v1_query_template_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Template executed \u2014 same shape as /query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UQLResponse"
                }
              }
            }
          },
          "404": {
            "description": "Template name not in registry"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResponseMeta": {
        "properties": {
          "data_as_of": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data As Of"
          },
          "record_count": {
            "type": "integer",
            "title": "Record Count",
            "default": 0
          },
          "query_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Ms"
          },
          "stale": {
            "type": "boolean",
            "title": "Stale",
            "default": false
          },
          "returned": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Returned"
          },
          "total_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Count"
          },
          "offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offset"
          },
          "limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "has_more": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Has More"
          },
          "next_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Offset"
          },
          "cache_hit": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Hit"
          },
          "includes_loaded": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Includes Loaded"
          },
          "staleness": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "fresh",
                  "stale",
                  "unknown"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Staleness"
          }
        },
        "type": "object",
        "title": "ResponseMeta",
        "description": "Response provenance + pagination meta.\n\nV1 fields (data_as_of/record_count/query_ms/stale) remain present and\ntyped identically for backward compatibility (spec \u00a7SC-003). The\nadditional fields below are populated by `services.uql.meta.build_meta`\nfor V2 UQL responses; V1 fixed-endpoint callers leave them as defaults."
      },
...
      "SortDirection": {
        "type": "string",
        "enum": [
          "asc",
          "desc"
        ],
        "title": "SortDirection"
      },
...
      "TemplateRequest": {
        "properties": {
          "template": {
            "type": "string",
            "minLength": 1,
            "title": "Template"
          },
          "params": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params"
          }
        },
        "type": "object",
        "required": [
          "template"
        ],
        "title": "TemplateRequest",
        "description": "Body for ``POST /api/v1/query/template``.\n\n``template`` is the registry key (see\n``backend/services/uql/templates.py::REGISTRY``); ``params`` is the\nbuilder-specific param dict (may be omitted for parameterless\ntemplates).",
        "examples": [
          {
            "params": {
              "limit": 5
            },
            "template": "top_rs_gainers"
          },
          {
            "params": {},
            "template": "sector_rotation"
          },
          {
            "params": {},
            "template": "breadth_dashboard"
          }
        ]
      },
...
      "UQLAggregation": {
        "properties": {
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "function": {
            "type": "string",
            "enum": [
              "avg",
              "sum",
              "min",
              "max",
              "count",
              "count_all",
              "pct_true",
              "pct_positive",
              "pct_above",
              "pct_below",
              "median",
              "stddev"
            ],
            "title": "Function"
          },
          "alias": {
            "type": "string",
            "minLength": 1,
            "title": "Alias"
          },
          "threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threshold"
          }
        },
        "type": "object",
        "required": [
          "function",
          "alias"
        ],
        "title": "UQLAggregation",
        "description": "One aggregation spec inside a UQL request (data-model \u00a71)."
      },
      "UQLFilter": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "op": {
            "$ref": "#/components/schemas/UQLOperator"
          },
          "value": {
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "field",
          "op"
        ],
        "title": "UQLFilter"
      },
      "UQLOperator": {
        "type": "string",
        "enum": [
          "=",
          "!=",
          ">",
          ">=",
          "<",
          "<=",
          "in",
          "not_in",
          "between",
          "contains",
          "is_null",
          "is_not_null"
        ],
        "title": "UQLOperator"
      },
      "UQLRequest": {
        "properties": {
          "entity_type": {
            "type": "string",
            "enum": [
              "equity",
              "mf",
              "sector",
              "index"
            ],
            "title": "Entity Type",
            "default": "equity"
          },
          "filters": {
            "items": {
              "$ref": "#/components/schemas/UQLFilter"
            },
            "type": "array",
            "title": "Filters",
            "default": []
          },
          "sort": {
            "items": {
              "$ref": "#/components/schemas/UQLSort"
            },
            "type": "array",
            "title": "Sort",
            "default": []
          },
          "group_by": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group By"
          },
          "aggregations": {
            "items": {
              "$ref": "#/components/schemas/UQLAggregation"
            },
            "type": "array",
            "title": "Aggregations",
            "default": []
          },
          "mode": {
            "type": "string",
            "enum": [
              "snapshot",
              "timeseries"
            ],
            "title": "Mode",
            "default": "snapshot"
          },
          "time_range": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UQLTimeRange"
              },
              {
                "type": "null"
              }
            ]
          },
          "granularity": {
            "type": "string",
            "const": "daily",
            "title": "Granularity",
            "default": "daily"
          },
          "fields": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields"
          },
          "include": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "enum": [
                    "identity",
                    "rs",
                    "technicals",
                    "conviction"
                  ]
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Include"
          },
          "limit": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 50
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset",
            "default": 0
          }
        },
        "type": "object",
        "title": "UQLRequest",
        "description": "UQL v2 request envelope. Mirrors specs/004-uql-aggregations contract.",
        "examples": [
          {
            "entity_type": "equity",
            "fields": [
              "symbol",
              "company_name",
              "rs_composite"
            ],
            "filters": [
              {
                "field": "sector",
                "op": "=",
                "value": "Banking"
              }
            ],
            "limit": 10,
            "sort": [
              {
                "direction": "desc",
                "field": "rs_composite"
              }
            ]
          },
          {
            "aggregations": [
              {
                "alias": "avg_rs",
                "field": "rs_composite",
                "function": "avg"
              },
              {
                "alias": "pct_above_200dma",
                "field": "above_200dma",
                "function": "pct_true"
              }
            ],
            "entity_type": "equity",
            "group_by": [
              "sector"
            ],
            "limit": 20
          },
          {
            "entity_type": "equity",
            "fields": [
              "date",
              "close",
              "rs_composite"
            ],
            "filters": [
              {
                "field": "symbol",
                "op": "=",
                "value": "RELIANCE"
              }
            ],
            "granularity": "daily",
            "limit": 100,
            "mode": "timeseries",
            "time_range": {
              "from": "2026-01-01",
              "to": "2026-04-10"
            }
          }
        ]
      },
      "UQLResponse": {
        "properties": {
          "records": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Records"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "type": "object",
        "required": [
          "records",
          "total",
          "meta"
        ],
        "title": "UQLResponse",
        "description": "Fully-shaped UQL response \u2014 records + total + provenance meta.",
        "examples": [
          {
            "meta": {
              "data_as_of": "2026-04-13",
              "has_more": false,
              "limit": 10,
              "offset": 0,
              "query_ms": 42,
              "returned": 1,
              "staleness": "fresh",
              "total_count": 1
            },
            "records": [
              {
                "company_name": "HDFC Bank",
                "rs_composite": 82.1,
                "symbol": "HDFCBANK"
              }
            ],
            "total": 1
          }
        ]
      },
      "UQLSort": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "direction": {
            "$ref": "#/components/schemas/SortDirection",
            "default": "desc"
          }
        },
        "type": "object",
        "required": [
          "field"
        ],
        "title": "UQLSort"
      },
      "UQLTimeRange": {
        "properties": {
          "from": {
            "type": "string",
            "format": "date",
            "title": "From"
          },
          "to": {
            "type": "string",
            "format": "date",
            "title": "To"
          }
        },
        "type": "object",
        "required": [
          "from",
          "to"
        ],
        "title": "UQLTimeRange",
        "description": "Inclusive date range for ``mode='timeseries'`` queries."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}